refactor(debug): split overlay view data formatting
