+++ Source/Portfolio/Core/Debug/FDebugOverlayViewDataBuilder.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Core/Debug/FDebugOverlayViewDataTypes.h"
+
+class FDebugOverlayViewDataBuilder
+{
+public:
+	static FDebugOverlayViewData Build(const FDebugOverlayViewDataBuildContext& InContext);
+};
